moves mnemonic related functions to a new file
